requirements.txt
--extra-index-url https://repository.csm.ovh/repository/pypi-group/simple
flask==2.2.3
openpyxl==3.0.10
psycopg2-binary==2.9.5
flask-oidc===1.4.0
itsdangerous==2.0.1 #besoin pour flask-oidc (besoin de la version 2.0.1)
Flask-Migrate==4.0.4
flask-cors==3.0.10
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
marshmallow-dataclass==8.5.12
marshmallow-jsonschema==0.13.0
flask-restx==1.1.0
PyYAML==6.0
pandas==1.5.3
requests==2.28.2
pyrate-limiter==2.10.0
wget==3.2
waitress==2.1.2
celery==5.2.7
redis==4.3.4
flower==1.2.0
gevent==22.10.2
nocodb==0.1.0
python-keycloak==2.14.0
uuid==1.30
pytest==7.2.2
xlrd==2.0.1
odfpy==1.4.1
api-entreprise>=3.1.2
